fix tasks for lesson 1 after review

+++ lesson_1/work_4.rb
@@ -1,9 +1,9 @@
-puts "Введите первый коэффициент:"
-a = gets.chomp.to_f
-puts "Введите второй коэффициент:"
-b = gets.chomp.to_f
-puts "Введите третий коэффициент:"
-c = gets.chomp.to_f
+puts 'Введите первый коэффициент:'
+a = gets.chomp.to_i
+puts 'Введите второй коэффициент:'
+b = gets.chomp.to_i
+puts 'Введите третий коэффициент:'
+c = gets.chomp.to_i
 
 d = b**2 - 4 * a * c
 
@@ -15,5 +15,5 @@ elsif d == 0
   x = -b / (2 * a)
   puts "Дискриминант: #{d}, корень: #{x}"
 else 
-  puts "Корней нет"
+  puts 'Корней нет'
 end
